+++ doczrc.js
@@ -1,0 +1,10 @@
+import { css } from 'docz-plugin-css'
+
+export default {
+  title: 'VTEX Butik',
+  plugins: [
+    css({
+      preprocessor: 'postcss',
+    }),
+  ],
+}